🚚 refactor: Change files naming convention to kebab-case
